changed cursor_position to size_t
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@2832 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ menu/text_control.h
@@ -27,7 +27,7 @@ private:
 	std::string _text, _value;
 	Alarm _blink;
 	bool _cursor_visible;
-	int _cursor_position;
+	size_t _cursor_position;
 };
 
 class HostTextControl : public TextControl {
